third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEBlendElement-dom-in-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' attribute of the SVGFEBlendElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS normalBlend.getAttribute('in') is "SourceAlpha"
PASS multiplyBlend.getAttribute('in') is "SourceAlpha"
PASS screenBlend.getAttribute('in') is "SourceAlpha"
PASS darkenBlend.getAttribute('in') is "SourceAlpha"
PASS lightenBlend.getAttribute('in') is "SourceAlpha"
PASS normalBlend.getAttribute('in') is "SourceGraphic"
PASS multiplyBlend.getAttribute('in') is "SourceGraphic"
PASS screenBlend.getAttribute('in') is "SourceGraphic"
PASS darkenBlend.getAttribute('in') is "SourceGraphic"
PASS lightenBlend.getAttribute('in') is "SourceGraphic"
PASS successfullyParsed is true

TEST COMPLETE
